+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+- 2019/01/22
+
+    - updated the `odttemplate` filter. It will use styles from the generated ODT file that haven't been present in the template file.
+
 - 2019/01/10
 
     - version 0.2d released